% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetZestimate.R
\name{GetZestimate}
\alias{GetZestimate}
\title{Make request to Zillow API GetZestimate Web Service}
\usage{
GetZestimate(
  zpid = NULL,
  rentzestimate = FALSE,
  zws_id = getOption("ZillowR-zws_id"),
  url = "http://www.zillow.com/webservice/GetZestimate.htm"
)
}
\arguments{
\item{zpid}{The Zillow Property ID for the property for which to obtain
information. Required.}

\item{rentzestimate}{Return Rent Zestimate information if available (logical,
default: false).}

\item{zws_id}{The Zillow Web Service Identifier. Required.}

\item{url}{URL for the GetZestimate Web Service. Required.}
}
\value{
A named list with the following elements:
\describe{
\item{\strong{request}}{a list with the request parameters}
\item{\strong{message}}{a list of status code(s) and message(s)
returned by the API}
\item{\strong{response}}{an XMLNode with the API-specific response
values. At this time, no further coercion is performed, so you
may have to use functions from the \code{XML} package to extract
the desired output.}
}
}
\description{
For a specified Zillow property identifier (zpid), the GetZestimate API
returns:
}
\details{
\itemize{
\item The most recent property Zestimate
\item The date the Zestimate was computed
\item The valuation range
\item The Zestimate ranking within the property's ZIP code.
\item The full property address and geographic location (latitude/longitude)
and a set of identifiers that uniquely represent the region (ZIP code,
city, county & state) in which the property exists.
}

The GetZestimate API will only surface properties for which a Zestimate
exists. If a request is made for a property that has no Zestimate, an error
code is returned. Zillow doesn't have Zestimates for all the homes in its
database. For such properties, we do have tax assessment data, but that is
not provided through the API. For more information, see our Zestimate
coverage.
}
\examples{
\dontrun{
GetZestimate(zpid = 48749425)
GetZestimate(zpid = 48749425, rentzestimate = TRUE)}
}
