% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZillowR-package.R
\docType{package}
\name{ZillowR-package}
\alias{ZillowR}
\alias{ZillowR-package}
\title{ZillowR: R Interface to Zillow Real Estate and Mortgage Data API}
\description{
Zillow, an online real estate company, provides real estate and
mortgage data for the United States through a REST API. The ZillowR package
provides an R function for each API service, making it easy to make API
calls and process the response into convenient, R-friendly data structures.
See <http://www.zillow.com/howto/api/APIOverview.htm> for the Zillow API
Documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/fascinatingfingers/zillowr}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/zillowr/issues}
}

}
\keyword{internal}
