% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZillowR-package.R
\docType{package}
\name{ZillowR-package}
\alias{ZillowR}
\alias{ZillowR-package}
\title{ZillowR: R Interface to Zillow Real Estate and Mortgage Data API}
\description{
Zillow, an online real estate company, provides real estate and mortgage data for the United States through a REST API. The ZillowR package provides an R function for each API service, making it easy to make API calls and process the response into convenient, R-friendly data structures. See <https://www.zillow.com/howto/api/APIOverview.htm> for the Zillow API Documentation. NOTE: Zillow deprecated their API on 2021-09-30, and this package is now deprecated as a result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fascinatingfingers.gitlab.io/zillowr}
  \item \url{https://gitlab.com/fascinatingfingers/zillowr}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/zillowr/-/issues}
}

}
\author{
\strong{Maintainer}: Justin Brantley \email{fascinatingfingers@icloud.com}

}
\keyword{internal}
